(feature) Adding NilUser
This makes life a lot easier for when there is no user. This person
can't do anything and **should** be laregely unused. However the layouts
are very much built expecting a user - the login is handled in a
different app.

Therefore adding this makes it really easy to run the app and un-pick
the layouts and controllers.